feat: refactor and added encryptionUtil testing
