+++ components/Vault/Open.vue
@@ -41,6 +41,7 @@
 </template>
 
 <script setup>
+import { decrypt, deriveKeyFromPassword } from '~/utils/encryptionUtils'
 definePageMeta({
   title: 'Open Vault',
   description: 'Open a vault',
@@ -96,119 +97,5 @@ async function openVault() {
       alert('Server error, try again later!')
     }
   }
-}
-
-async function deriveKeyFromPassword(password) {
-  const salt = new Uint8Array([1, 2, 3, 4])
-  const encoder = new TextEncoder()
-  const encodedPassword = encoder.encode(password)
-
-  // Import key here is used to set the "structure" of the key
-  const derivedKey = await crypto.subtle.importKey(
-    'raw',
-    encodedPassword,
-    { name: 'PBKDF2' },
-    false,
-    ['deriveKey']
-  )
-  const key = await crypto.subtle.deriveKey(
-    {
-      name: 'PBKDF2',
-      salt: salt,
-      iterations: 110000,
-      hash: 'SHA-256',
-    },
-    derivedKey,
-    { name: 'AES-GCM', length: 256 },
-    false,
-    ['encrypt', 'decrypt']
-  )
-
-  return key
-}
-
-// Encryption decryption functions
-async function encrypt(stringToEncrypt, encryptionKeyObject) {
-  const encoder = new TextEncoder()
-  const encodedString = encoder.encode(stringToEncrypt)
-
-  const iv = crypto.getRandomValues(new Uint8Array(12))
-  const encryptedData = await crypto.subtle.encrypt(
-    { name: 'AES-GCM', iv },
-    encryptionKeyObject,
-    encodedString
-  )
-
-  const stringArray = new Uint8Array(encryptedData)
-  const base64String = toBase64Url(stringArray)
-  const base64IV = toBase64Url(iv)
-
-  return base64IV + base64String
-}
-
-async function decrypt(stringToDecrypt, encryptionKeyObject) {
-  const base64IV = stringToDecrypt.slice(0, 16)
-  const base64String = stringToDecrypt.slice(16)
-
-  const encStringUInt8Array = fromBase64Url(base64String)
-  const encryptedString = encStringUInt8Array.buffer
-
-  const ivUInt8Array = fromBase64Url(base64IV)
-  const iv = ivUInt8Array.buffer
-
-  const decryptedData = await crypto.subtle.decrypt(
-    { name: 'AES-GCM', iv },
-    encryptionKeyObject,
-    encryptedString
-  )
-
-  const decryptedString = new TextDecoder().decode(decryptedData)
-
-  return decryptedString
-}
-
-function toBase64Url(byteArray) {
-  // Convert byteArray to a standard base64 string
-  const base64String = window.btoa(String.fromCharCode.apply(null, byteArray))
-
-  // Make the base64 string URL and filename safe
-  const base64UrlString = base64String
-    .replace(/\+/g, '-')
-    .replace(/\//g, '_')
-    .replace(/=+$/, '')
-
-  return base64UrlString
-}
-
-function fromBase64Url(base64UrlString) {
-  // Replace URL-safe characters back to their original
-  const base64String = base64UrlString.replace(/-/g, '+').replace(/_/g, '/')
-
-  // Pad the base64 string to make its length a multiple of 4
-  const paddedBase64String = base64String.padEnd(
-    base64String.length + ((4 - (base64String.length % 4)) % 4),
-    '='
-  )
-
-  // Decode base64 string to a UTF-16 string
-  const decodedString = window.atob(paddedBase64String)
-
-  // Convert decoded string to byte array
-  const byteArray = new Uint8Array(decodedString.length)
-  for (let i = 0; i < decodedString.length; i++) {
-    byteArray[i] = decodedString.charCodeAt(i)
-  }
-
-  return byteArray
-}
-
-function base64ToArrayBuffer(base64) {
-  const binaryString = atob(base64)
-  const len = binaryString.length
-  const bytes = new Uint8Array(len)
-  for (let i = 0; i < len; i++) {
-    bytes[i] = binaryString.charCodeAt(i)
-  }
-  return bytes.buff
 }
 </script>
